feature: add WordInput Protocol
